:sparkles: feat: Add request body parsing and file writing
This commit introduces two main changes to `server.go`. Firstly, the request parsing logic has been updated to include the body of the request. A `Body` field has been added to the `request` struct, and the parsing loop now includes a condition to start reading the body once the headers have been parsed.

Secondly, the `fileCreation` function has been updated to write the body of the request to a file. The URI is cleaned to remove the "/files/" prefix, and then the body of the request is written to a file at the specified directory. If the file writing fails, an error message is printed.